change menu coloumn

+++ database/migrations/2024_12_27_074428_create_menus_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->integer('category_id');
             $table->string('menu_img')->nullable();
             $table->text('menu_description')->nullable();
-            $table->integer('menu_s_price')->nullable();
-            $table->integer('menu_l_price')->nullable();
+            $table->double('menu_s_price')->nullable();
+            $table->double('menu_l_price')->nullable();
             $table->integer('menu_status')->default(1);
             $table->timestamps();
         });
